+++ src/Uri.cpp
@@ -473,7 +473,7 @@ std::string Uri::extension() const
         return "";
 
     auto parts = Luau::split(components.back(), '.');
-    if (parts.empty())
+    if (parts.size() == 1)
         return "";
 
     return "." + std::string(parts.back());